requirements.txt
ollama
faiss-cpu
numpy
langchain
sentence-transformers
langchain_community
pypdf